Update unit tests framework per the requested changes
- Renamed utest to unit_test
- Moved test supervisor to the examples/services directory
- Moved unit tests to examples/unit_tests
- Made a README for the unit test framework
- Made READMEs for the individual unit tests
- No longer a separate standalone library, but instead integrated with libtock
