+++ userland/examples/services/unit_test_supervisor/Makefile
@@ -1,10 +1,9 @@
 # Makefile for user application
 
+PACKAGE_NAME="org.tockos.unit_test"
+
 # Specify this directory relative to the current application.
 TOCK_USERLAND_BASE_DIR = ../../..
-
-# External libraries used
-EXTERN_LIBS += $(TOCK_USERLAND_BASE_DIR)/libutest
 
 # Which files to compile.
 C_SRCS := $(wildcard *.c)
