+++ userland/examples/services/unit_test_supervisor/main.c
@@ -1,0 +1,6 @@
+#include <unit_test.h>
+
+int main(void) {
+  unit_test_service();
+  return 0;
+}
