+++ userland/examples/unit_tests/passfail/Makefile
@@ -1,12 +1,7 @@
 # Makefile for user application
 
-PACKAGE_NAME="org.tockos.utest"
-
 # Specify this directory relative to the current application.
 TOCK_USERLAND_BASE_DIR = ../../..
-
-# External libraries used
-EXTERN_LIBS += $(TOCK_USERLAND_BASE_DIR)/libutest
 
 # Which files to compile.
 C_SRCS := $(wildcard *.c)
