--- userland/examples/utest/utest_supervisor/main.c
@@ -1,6 +1,0 @@
-#include <utest.h>
-
-int main(void) {
-    utest_service();
-    return 0;
-}
