--- userland/libutest/Makefile
@@ -1,11 +1,0 @@
-TOCK_USERLAND_BASE_DIR ?= ..
-LIBNAME := libutest
-$(LIBNAME)_DIR := $(TOCK_USERLAND_BASE_DIR)/$(LIBNAME)
-
-# List all C and Assembly files
-$(LIBNAME)_SRCS  := $(wildcard $($(LIBNAME)_DIR)/internal/*.c) $(wildcard $($(LIBNAME)_DIR)/*.c) $(wildcard $($(LIBNAME)_DIR)/*.s)
-
-# Need libtock headers
-override CPPFLAGS += -I$(TOCK_USERLAND_BASE_DIR)/libtock
-
-include $(TOCK_USERLAND_BASE_DIR)/TockLibrary.mk
